+++ packages/vue/src/components/select/select-root.vue
@@ -1,29 +1,29 @@
 <script lang="ts">
 import type { HTMLAttributes } from 'vue'
-import type { BooleanDefaults } from '../../types'
+import type { BooleanDefaults, CollectionItem } from '../../types'
 import type { RenderStrategyProps } from '../../utils'
 import type { PolymorphicProps } from '../factory'
 import type { RootEmits, RootProps } from './select.types'
 
-export interface SelectRootProps
-  extends RootProps,
+export interface SelectRootProps<T extends CollectionItem>
+  extends RootProps<T>,
     RenderStrategyProps,
     PolymorphicProps,
     /**
      * @vue-ignore
      */
     HTMLAttributes {}
-export interface SelectRootEmits extends RootEmits {}
+export type { RootEmits as SelectRootEmits } from './select.types'
 </script>
 
-<script setup lang="ts">
+<script setup lang="ts" generic="T extends CollectionItem">
 import { computed } from 'vue'
 import { RenderStrategyPropsProvider } from '../../utils'
 import { ark } from '../factory'
 import { useSelect } from './use-select'
 import { SelectProvider } from './use-select-context'
 
-const props = withDefaults(defineProps<SelectRootProps>(), {
+const props = withDefaults(defineProps<SelectRootProps<T>>(), {
   closeOnSelect: undefined,
   composite: undefined,
   defaultOpen: undefined,
@@ -33,9 +33,9 @@ const props = withDefaults(defineProps<SelectRootProps>(), {
   multiple: undefined,
   open: undefined,
   readOnly: undefined,
-} satisfies BooleanDefaults<RootProps>)
+} satisfies BooleanDefaults<RootProps<T>>)
 
-const emits = defineEmits<SelectRootEmits>()
+const emits = defineEmits<RootEmits<T>>()
 
 const select = useSelect(props, emits)
 SelectProvider(select)